+++ src/Bpp/Phyl/Io/BppOSubstitutionModelFormat.cpp
@@ -1201,14 +1201,14 @@ SubstitutionModel* BppOSubstitutionModelFormat::readWord_(const Alphabet* alphab
         name+="AAClust";
 
         // Initialization using the "assign" argument
-        vector<uint> partition;
+        vector<unsigned int> partition;
         if (args.find("partition")!=args.end())
         {
           string rf = args["partition"];
 
           StringTokenizer strtok(rf.substr(1, rf.length() - 2), ",");
           while (strtok.hasMoreToken())
-            partition.push_back(TextTools::to<uint>(strtok.nextToken()));
+            partition.push_back(TextTools::to<unsigned int>(strtok.nextToken()));
         }
 
         AbstractCodonClusterAASubstitutionModel* aca=partition.size()!=0?
@@ -1708,7 +1708,7 @@ void BppOSubstitutionModelFormat::write(const TransitionModel& model,
   const Coala* coalaModel = dynamic_cast<const Coala*>(&model);
   if (coalaModel)
   {
-    out << "exch=" << coalaModel->getExch() << ",nbrAxes=" << coalaModel->getNbrOfAxes();
+    out << "exch=" << coalaModel->getExch() << ",nbrAxes=" << (int)coalaModel->getNbrOfAxes();
     comma = true;
   }
 
@@ -1771,7 +1771,7 @@ void BppOSubstitutionModelFormat::write(const TransitionModel& model,
         if (comma)
           out << ",";
         out << "partition=(";
-        const vector<uint>& vass=acc->getAssign();
+        const vector<unsigned int>& vass=acc->getAssign();
         
         for (size_t j=0; j<vass.size(); j++)
         {
@@ -1806,7 +1806,7 @@ void BppOSubstitutionModelFormat::write(const TransitionModel& model,
   {
     if (comma)
       out << ",";
-    out << "n=" << pM7->getNumberOfModels();
+    out << "n=" << (int) pM7->getNumberOfModels();
     comma=true;
   }
 
@@ -1815,7 +1815,7 @@ void BppOSubstitutionModelFormat::write(const TransitionModel& model,
   {
     if (comma)
       out << ",";
-    out << "n=" << pM8->getNumberOfModels()-1;
+    out << "n=" << (int) pM8->getNumberOfModels()-1;
     comma=true;
   }
